+++ R/project.R
@@ -100,14 +100,15 @@ link-citations: yes", output_theme, '
 
 ```{r setup, include = FALSE}
 path <- "', getwd(), '"
-knitr::opts_chunk$set(eval=TRUE, #run code in chunks (default = TRUE)
-                       echo = TRUE, #whether to include the source code in the output
-                       error = TRUE, #Whether to include error information in the output
-                       warning = FALSE, #Whether to include warnings in the output (default = TRUE)
-                       message = TRUE, #whether to include reference information in the output
-                       cache=TRUE, #whether to cache
-                       collapse = FALSE # output in one piece
-                       )
+knitr::opts_chunk$set(
+  eval = TRUE, # run code in chunks (default = TRUE)
+  echo = TRUE, # whether to include the source code in the output
+  error = TRUE, # Whether to include error information in the output
+  warning = FALSE, # Whether to include warnings in the output (default = TRUE)
+  message = TRUE, # whether to include reference information in the output
+  cache = TRUE, # whether to cache
+  collapse = FALSE # output in one piece
+)
 knitr::opts_knit$set(root.dir = path)
 ```
 
@@ -198,6 +199,65 @@ make_asa_web <- function() {
   system("quarto render ~/Documents/GitHub/Asa_web/ --quiet")
   # rmarkdown::render_site(encoding = "UTF-8", input = "~/Documents/GitHub/Asa_web/", quiet = TRUE)
   system("rsync -a ~/Documents/GitHub/Asa_web/_site/* ~/Documents/GitHub/Asa12138.github.io/")
+}
+
+add_common_gitignore <- function(dir = getwd(), rm_DS_Store = TRUE) {
+  if (!dir.exists(dir)) stop("directory not exist")
+  old_dir <- getwd()
+  on.exit(setwd(old_dir))
+  setwd(dir)
+  if (!dir.exists(".git")) stop("not a git repository")
+  common_r <- "# Common R files
+.Rproj.user
+.Rhistory
+.RData
+.Ruserdata
+.Renviron
+*.Rproj
+.quarto
+.quarto/
+.DS_Store
+**/.DS_Store
+.DS_Store?
+*_cache/
+/cache/
+# R books
+_bookdown_files
+_book
+_freeze
+# pkgdown site
+docs/
+# RStudio Connect folder
+rsconnect/
+"
+  # 读取.gitignore文件已有内容
+  if (file.exists(".gitignore")) {
+    readLines(".gitignore") -> tmp
+    if (any(grepl("# Common R files", tmp))) {
+      message("already have common R files")
+    } else {
+      writeLines(paste0(common_r, paste0(tmp, collapse = "\n"), "\n"), ".gitignore")
+    }
+  } else {
+    writeLines(common_r, ".gitignore")
+  }
+
+  if (rm_DS_Store) system("find . -name .DS_Store -print0 | xargs -0 git rm -f --ignore-unmatch")
+}
+
+publish_book2github <- function(dir, overwrite = FALSE) {
+  if (!dir.exists(dir)) stop("directory not exist")
+  old_dir <- getwd()
+  on.exit(setwd(old_dir))
+  setwd(dir)
+  if (!dir.exists("_book")) stop("not a book repository")
+  # rmarkdown::render_site(encoding = "UTF-8", input = "~/Documents/GitHub/Asa_web/", quiet = TRUE)
+  destination <- paste0("~/Documents/GitHub/Asa12138.github.io/", basename(dir))
+  if (overwrite) {
+    system(paste0("rm -rf ", destination))
+    system(paste0("cp -R _book ", destination))
+  }
+  system(paste0("rsync -a _book/* ", destination))
 }
 
 # =======Some tips========
@@ -354,10 +414,13 @@ for (( N = $START; N <= $STOP; N++ ))
 do
   sample=$(head -n \"$N\" $samplelist | tail -n 1)
 	echo $sample
+	start1=`date +%s`
 
   bowtie2 -p 8 -x ~/db/humangenome/hg38 -1 ~/work/st/data/fastp/${sample}_1.fq -2 ~/work/st/data/fastp/${sample}_2.fq \
   -S ~/work/st/data/rm_human/${sample}.sam --un-conc ~/work/st/data/rm_human/${sample}.fq --very-sensitive
 
+	end1=`date +%s`
+  echo `expr $end1 - $start1`s
 done
 ##############
 echo end: `date +'%Y-%m-%d %T'`
@@ -521,7 +584,6 @@ update_NEWS_md <- function(package_dir = ".", new_features = character(),
   pkg_name <- pkg_info$Package
   new_version <- pkg_info$Version
 
-  writeLines(".*\\.md", file.path(package_dir, ".Rbuildignore"))
   file_path <- file.path(package_dir, "NEWS.md")
   # Parse existing NEW.md content
   if (file.exists(file_path)) {
@@ -596,7 +658,7 @@ get_package_info <- function(package_dir = ".") {
 
 solve_no_visible_binding <- function(file = "~/Downloads/00check.log.txt") {
   # 读入文件
-  log_file <- readLines("~/Downloads/00check.log.txt")
+  log_file <- readLines(file)
 
   # 挑出带有 'no visible binding for global variable' 的行
   error_lines <- grep("no visible binding for global variable", log_file, value = TRUE)
@@ -612,17 +674,18 @@ solve_no_visible_binding <- function(file = "~/Downloads/00check.log.txt") {
 
   # 初始化临时变量
   tmp <- ""
-  tmp2 <- ""
+  tmp2 <- c()
 
   # 循环处理错误信息
   for (msg in error_messages) {
     x <- strsplit(msg, ": ")[[1]]
 
     if (tmp == x[1]) {
-      tmp2 <- paste0(tmp2, "=", x[2])
+      # tmp2 <- paste0(tmp2, " <- ", x[2])
+      tmp2 <- c(tmp2, x[2])
     } else {
       if (tmp != "") {
-        output <- paste(output, paste0(tmp, ":", tmp2, "=NULL\n"))
+        output <- paste(output, paste0(tmp, ": ", paste0(unique(tmp2), collapse = " <- "), " <- NULL\n"))
       }
       tmp <- x[1]
       tmp2 <- x[2]
@@ -630,7 +693,7 @@ solve_no_visible_binding <- function(file = "~/Downloads/00check.log.txt") {
   }
 
   # 处理最后一个
-  output <- paste(output, paste0(tmp, ":", tmp2, "=NULL"))
+  output <- paste(output, paste0(tmp, ": ", paste0(unique(tmp2), collapse = " <- "), " <- NULL"))
 
   # 输出结果
   cat(output, sep = "\n")
@@ -745,20 +808,26 @@ check_print_cat_in_R_files <- function(package_folder_path = ".", exclude = "pri
 #' @param exclude vector for excluding .R files
 #' @param indent_by indent_by, default: 2
 #' @param ... other parameters for devtools::check
+#' @param check check or not, default: TRUE
 #'
 #' @return No value
 #' @export
-prepare_package <- function(pkg_dir = ".", exclude = "print.R", indent_by = 2, ...) {
+prepare_package <- function(pkg_dir = ".", exclude = "print.R", indent_by = 2, check = TRUE, ...) {
   if (!interactive()) stop("This function is only for interactive use")
   # Check the package name is available or not
   # available::available(get_package_info(pkg_dir)$Package)
   pcutils::dabiao("Styler all codes")
   styler::style_pkg(pkg = pkg_dir, strict = TRUE, indent_by = indent_by)
-  pcutils::dabiao("Lint all codes")
-  my_linters <- lintr::linters_with_defaults(
-    indentation_linter = lintr::indentation_linter(indent = indent_by)
-  )
-  lintr::lint_package(path = pkg_dir, linters = my_linters)
+  if (FALSE) {
+    pcutils::dabiao("Lint all codes")
+    my_linters <- lintr::linters_with_defaults(
+      indentation_linter = lintr::indentation_linter(indent = indent_by)
+    )
+    lintr::lint_package(path = pkg_dir, linters = my_linters)
+  } else {
+    message("Skip linting, do `lintr::lint_package` yourself.")
+  }
+
   pcutils::dabiao("Write documents")
   devtools::document(pkg_dir)
   check_Rds(pkg_dir)
@@ -766,14 +835,27 @@ prepare_package <- function(pkg_dir = ".", exclude = "print.R", indent_by = 2, .
   check_TF_in_R_files(pkg_dir)
   pcutils::dabiao("Check cat/print")
   check_print_cat_in_R_files(pkg_dir, exclude = exclude)
-  pcutils::dabiao("Check")
-  devtools::check(pkg_dir, ...)
+  if (check) {
+    pcutils::dabiao("Check")
+    devtools::check(pkg_dir, ...)
+  }
 }
 
 #' Re-install my packages
+#'
+#' @param are_you_pc are_you_pc? default: TRUE
 #' @param pkgs pkgs
+#'
 #' @return No return value
-reinstall_my_packages <- function(pkgs = c("pcutils", "pctax", "MetaNet", "ReporterScore")) {
+#' @noRd
+reinstall_my_packages <- function(pkgs = c("pcutils", "pctax", "MetaNet", "ReporterScore"), are_you_pc = TRUE) {
+  if (!are_you_pc) {
+    for (i in pkgs) {
+      devtools::install_github(paste0("Asa12138/", i))
+    }
+    return(invisible())
+  }
+
   for (i in pkgs) {
     if (i == "pctax") i <- "pctax/pctax"
     if (i == "MetaNet") i <- "MetaNet/MetaNet"
@@ -790,4 +872,79 @@ reinstall_my_packages <- function(pkgs = c("pcutils", "pctax", "MetaNet", "Repor
       }
     )
   }
+}
+
+#' Build CRAN version package
+#'
+#' @param pkg_dir pkg_dir
+#' @param additional additional
+#' @param check check, default: TRUE
+#'
+#' @return No value
+#' @noRd
+build_cran_pkg <- function(pkg_dir = "./", additional = "R/additional.R", check = TRUE) {
+  pkg_dir <- normalizePath(pkg_dir)
+
+  # 先复制一份pkg_dir到临时目录
+  tmp_dir <- tempdir()
+  temp_pkg_dir <- file.path(tmp_dir, basename(pkg_dir))
+
+  # 检查是不是R包目录
+  if (!file.exists(file.path(pkg_dir, "DESCRIPTION"))) {
+    stop("It's not a R package directory!")
+  }
+
+  if (dir.exists(temp_pkg_dir)) {
+    unlink(temp_pkg_dir, recursive = TRUE)
+  }
+  file.copy(from = pkg_dir, to = tmp_dir, recursive = TRUE, overwrite = TRUE)
+  # 进入临时目录
+  old_dir <- getwd()
+  on.exit(setwd(old_dir))
+  setwd(temp_pkg_dir)
+
+  # 读取DESCRIPTION文件
+  desc_file <- file.path(temp_pkg_dir, "DESCRIPTION")
+  desc <- read.dcf(desc_file) %>% as.data.frame()
+  sugg_pkgs <- strsplit(desc$Suggests, ",")[[1]] %>% trimws()
+
+  # 获取additional文件中的## some suggested pkgs: start到## some suggested pkgs: end中间部分的包
+  if (!is.null(additional)) {
+    if (file.exists(additional)) {
+      additional_content <- readLines(additional)
+      additional_content <- additional_content[(which(grepl("## some suggested pkgs: start", additional_content)) + 1):(which(grepl("## some suggested pkgs: end", additional_content)) - 1)]
+      # 然后在desc中删除这部分additional_content
+      gsub("#", "", additional_content) %>%
+        gsub(",$", "", .) %>%
+        strsplit(",") %>%
+        unlist() %>%
+        trimws() -> additional_pkgs
+      message("additional_pkgs: ", paste0(additional_pkgs, collapse = ", "))
+      sugg_pkgs <- setdiff(sugg_pkgs, additional_pkgs)
+      desc$Suggests <- paste0(sugg_pkgs, collapse = ",\n")
+      write.dcf(desc, desc_file)
+    }
+  }
+
+  # 删除additional
+  if (!is.null(additional)) {
+    if (file.exists(additional)) {
+      file.remove(additional)
+    }
+  }
+
+  # 删除NAMESPACE文件
+  file.remove(file.path(temp_pkg_dir, "NAMESPACE"))
+
+  # 重新检查
+  prepare_package(check = check)
+  # 重新构建
+  system(paste0("R CMD build ../", basename(pkg_dir)))
+  # 复制到原目录下
+  pkg <- dir(file.path(temp_pkg_dir), pattern = "*.tar.gz", full.names = TRUE)
+  file.copy(pkg, old_dir)
+  # 删除临时目录
+  setwd(old_dir)
+  unlink(temp_pkg_dir, recursive = TRUE)
+  message("The CRAN version package has been built successfully! at ", old_dir)
 }